+++ src/Plugins/Win32ConsoleLogger/Win32ConsoleLogger.cpp
@@ -108,7 +108,7 @@ namespace Mengine
         m_createConsole = false;
     }
     //////////////////////////////////////////////////////////////////////////
-    void Win32ConsoleLogger::log( ELoggerLevel _level, uint32_t _flag, const Char * _data, uint32_t _count )
+    void Win32ConsoleLogger::log( ELoggerLevel _level, uint32_t _flag, const Char * _data, size_t _count )
     {
         MENGINE_UNUSED( _flag );
